ported r81039 [type cast in sock_write()] from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-4-1-branch@81040 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/internet/sockconn.c
@@ -188,7 +188,7 @@ static size_t sock_write(const void *ptr, size_t size, size_t nitems,
 			 Rconnection con)
 {
     Rsockconn this = (Rsockconn)con->private;
-    ssize_t n = R_SockWrite(this->fd, ptr, (int)(size * nitems),
+    ssize_t n = R_SockWrite(this->fd, ptr, (size_t)(size * nitems),
 			    this->timeout)/((ssize_t)size);
     return n > 0 ? n : 0;
 }
